# [v1.8][cherry-pick][quorum store][mempool] optimize pulling txns for when many txns are in progress (#10744)