@sherry-x 2023-11-14
for testing, I assume the existing test not always capture this issue (or else we'd find it earlier?) do we need to add more test to verify this? how did we reproduced and verified?