@bchocho 2023-11-14
@sherry-x good question -- I should have made the description more clear (which is already in main git history).

We know that the conditions of the issue triggering was (i) pull takes > 25ms and (ii) there are no new transactions in the pull. Because of (ii) this is pretty rare and like you said it wouldn't trigger all the time.

(1) the pull optimization was measured

In the [new test run](https://aptoslabs.grafana.net/d/consensus/consensus?from=1699567187000&to=1699568085000&var-Datasource=fHo-R604z&var-BigQuery=axNEitxVz&var-metrics_source=All&var-chain_name=forge-1&var-cluster=All&var-namespace=forge-runner-1699567182&var-kubernetes_pod_name=All&orgId=1&viewPanel=86), the pull is comfortably below (i) at < 8 ms where previously it would go above 25 ms.

(2) the logic was changed so the behavior when hitting (ii) can no longer trigger something like this
(a) we removed the biased in the tokio select
(b) we explicitly check whether the pull takes longer than the tick time (> 25 ms) and back off if that is the case